message

+++ database/cartcode.php
@@ -5,17 +5,15 @@ if(isset($_GET["action"]))
 {
 if($_GET["action"] == "delete")
 {
-foreach($_SESSION["cart"] as $keys => $values)
-{
-if($values["food_id"] == $_GET["id"])
-{
-unset($_SESSION["cart"][$keys]);
-echo '<script>alert("Food has been removed")</script>';
-echo '<script>window.location="../cart.php"</script>';
+if(array_key_exists('cart', $_SESSION) AND !empty($_SESSION['cart'])){
+    unset($_SESSION['cart'][$_GET['id']]);
+    echo '<script>alert("Food has been removed")</script>';
+    echo '<script>window.location="../customer/cart.php"</script>';
 }
 }
 }
-}
+
+
 
 
 if(isset($_GET["action"]))
@@ -26,7 +24,7 @@ foreach($_SESSION["cart"] as $keys => $values)
 {
 unset($_SESSION["cart"]);
 echo '<script>alert("Cart is made empty!")</script>';
-echo '<script>window.location="../cart.php"</script>';
+echo '<script>window.location="../customer/cart.php"</script>';
 }
 }
 }
